APIs/azure.com/resources-locks/2016-09-01/python-aiohttp-client/tests/test_management_locks_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.management_lock_list_result import ManagementLockListResult
from openapi_server.models.management_lock_object import ManagementLockObject


pytestmark = pytest.mark.asyncio

async def test_management_locks_create_or_update_at_resource_group_level(client):
    """Test case for management_locks_create_or_update_at_resource_group_level

    Creates or updates a management lock at the resource group level.
    """
    parameters = {"name":"name","id":"id","type":"type","properties":{"notes":"notes","level":"NotSpecified","owners":[{"applicationId":"applicationId"},{"applicationId":"applicationId"}]}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PUT',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Authorization/locks/{lock_name}'.format(resource_group_name='resource_group_name_example', lock_name='lock_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_create_or_update_at_resource_level(client):
    """Test case for management_locks_create_or_update_at_resource_level

    Creates or updates a management lock at the resource level or any level below the resource.
    """
    parameters = {"name":"name","id":"id","type":"type","properties":{"notes":"notes","level":"NotSpecified","owners":[{"applicationId":"applicationId"},{"applicationId":"applicationId"}]}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PUT',
        path='/subscriptions/{subscription_id}/resourcegroups/{resource_group_name}/providers/{resource_provider_namespace}/{parent_resource_path}/{resource_type}/{resource_name}/providers/Microsoft.Authorization/locks/{lock_name}'.format(resource_group_name='resource_group_name_example', resource_provider_namespace='resource_provider_namespace_example', parent_resource_path='parent_resource_path_example', resource_type='resource_type_example', resource_name='resource_name_example', lock_name='lock_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_create_or_update_at_subscription_level(client):
    """Test case for management_locks_create_or_update_at_subscription_level

    Creates or updates a management lock at the subscription level.
    """
    parameters = {"name":"name","id":"id","type":"type","properties":{"notes":"notes","level":"NotSpecified","owners":[{"applicationId":"applicationId"},{"applicationId":"applicationId"}]}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PUT',
        path='/subscriptions/{subscription_id}/providers/Microsoft.Authorization/locks/{lock_name}'.format(lock_name='lock_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_create_or_update_by_scope(client):
    """Test case for management_locks_create_or_update_by_scope

    
    """
    parameters = {"name":"name","id":"id","type":"type","properties":{"notes":"notes","level":"NotSpecified","owners":[{"applicationId":"applicationId"},{"applicationId":"applicationId"}]}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PUT',
        path='/{scope}/providers/Microsoft.Authorization/locks/{lock_name}'.format(scope='scope_example', lock_name='lock_name_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_delete_at_resource_group_level(client):
    """Test case for management_locks_delete_at_resource_group_level

    Deletes a management lock at the resource group level.
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Authorization/locks/{lock_name}'.format(resource_group_name='resource_group_name_example', lock_name='lock_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_delete_at_resource_level(client):
    """Test case for management_locks_delete_at_resource_level

    Deletes the management lock of a resource or any level below the resource.
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/subscriptions/{subscription_id}/resourcegroups/{resource_group_name}/providers/{resource_provider_namespace}/{parent_resource_path}/{resource_type}/{resource_name}/providers/Microsoft.Authorization/locks/{lock_name}'.format(resource_group_name='resource_group_name_example', resource_provider_namespace='resource_provider_namespace_example', parent_resource_path='parent_resource_path_example', resource_type='resource_type_example', resource_name='resource_name_example', lock_name='lock_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_delete_at_subscription_level(client):
    """Test case for management_locks_delete_at_subscription_level

    Deletes the management lock at the subscription level.
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/subscriptions/{subscription_id}/providers/Microsoft.Authorization/locks/{lock_name}'.format(lock_name='lock_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_delete_by_scope(client):
    """Test case for management_locks_delete_by_scope

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/{scope}/providers/Microsoft.Authorization/locks/{lock_name}'.format(scope='scope_example', lock_name='lock_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_get_at_resource_group_level(client):
    """Test case for management_locks_get_at_resource_group_level

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Authorization/locks/{lock_name}'.format(resource_group_name='resource_group_name_example', lock_name='lock_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_get_at_resource_level(client):
    """Test case for management_locks_get_at_resource_level

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourcegroups/{resource_group_name}/providers/{resource_provider_namespace}/{parent_resource_path}/{resource_type}/{resource_name}/providers/Microsoft.Authorization/locks/{lock_name}'.format(resource_group_name='resource_group_name_example', resource_provider_namespace='resource_provider_namespace_example', parent_resource_path='parent_resource_path_example', resource_type='resource_type_example', resource_name='resource_name_example', lock_name='lock_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_get_at_subscription_level(client):
    """Test case for management_locks_get_at_subscription_level

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/providers/Microsoft.Authorization/locks/{lock_name}'.format(lock_name='lock_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_get_by_scope(client):
    """Test case for management_locks_get_by_scope

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/{scope}/providers/Microsoft.Authorization/locks/{lock_name}'.format(scope='scope_example', lock_name='lock_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_list_at_resource_group_level(client):
    """Test case for management_locks_list_at_resource_group_level

    
    """
    params = [('$filter', 'filter_example'),
                    ('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Authorization/locks'.format(resource_group_name='resource_group_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_list_at_resource_level(client):
    """Test case for management_locks_list_at_resource_level

    
    """
    params = [('$filter', 'filter_example'),
                    ('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourcegroups/{resource_group_name}/providers/{resource_provider_namespace}/{parent_resource_path}/{resource_type}/{resource_name}/providers/Microsoft.Authorization/locks'.format(resource_group_name='resource_group_name_example', resource_provider_namespace='resource_provider_namespace_example', parent_resource_path='parent_resource_path_example', resource_type='resource_type_example', resource_name='resource_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_list_at_subscription_level(client):
    """Test case for management_locks_list_at_subscription_level

    
    """
    params = [('$filter', 'filter_example'),
                    ('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/providers/Microsoft.Authorization/locks'.format(subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_management_locks_list_by_scope(client):
    """Test case for management_locks_list_by_scope

    
    """
    params = [('$filter', 'filter_example'),
                    ('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/{scope}/providers/Microsoft.Authorization/locks'.format(scope='scope_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

